requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.6
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.6.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
